nix: build yara
yara is a dependency of smartvmi's inmemoryscanner plugin. yara is
available in nixpkgs, but it is not a CMake project. Thus, the yara
package from nixpkgs does not provided a ...config.cmake file, which is
required, so that CMake can find yara when building smartvmi.
This commit overrides the original yara package and adds a
yara-config.cmake file.