smartvmi: add gtest dependency to build test successfully
previously, the Nix packaging from this repository assumed that the
CMakeLists.txt are modified in such a way that the tests are not built
by default.
https://github.com/GDATASoftwareAG/smartvmi/pull/154#discussion_r2073328561
requested to remove the behavior. The suggestion to build the
CMakeLists.txt from the src folder instead of the top level is great,
but after a few tries I did not find the correct derivation settings
(e.g., by setting cmakeDir) and chose to just build the tests as well.
This commit switches the smartvmi source to a commit without the
CMakeLists.txt changes and adds gtest to both vmicore and the plugins
(required dependency of the test code).